requirements.txt
crawl4ai==0.4.247
python-wordpress-xmlrpc==2.3
beautifulsoup4==4.12.3
requests==2.31.0
playwright>=1.49.0
rich==13.7.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6